pyproject.toml
[project]
name = "siamese-image-diff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
